First commit of search UX rework: Added SearchMenu and SettingsMenu, removed SecondaryMenu, reworked SearchResultsMenu to provide a menu section to SearchMenu and changed wiring and naming throughout the code base. Also did a general tidy up of files around the menu systems.
Buddy: Tom

+++ ios/ios_src/Menu/View/CustomTableViewCell.mm
@@ -91,7 +91,7 @@ const float SubViewInset = 22.f;
 
     if(!m_rightAlignImage)
     {
-        const float delta = (imageFrame.origin.x - initialImageX);
+        const float delta = static_cast<float>(imageFrame.origin.x - initialImageX);
 
         CGRect labelFrame = self.textLabel.frame;
         labelFrame.origin.x += delta;
